refactor(preact): update calendar and current months
