+++ src/components/Calendar/cells/WeekDay.jsx
@@ -5,7 +5,7 @@ This Source Code Form is subject to the terms of the Mozilla Public License, v. 
 the MPL was not distributed with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 */
 
-import style from "../../styles/calendar.module.css";
+import style from "../style.module.css";
 
 const longFormat = new Intl.DateTimeFormat("de-DE", { weekday: "long" });
 const shortFormat = new Intl.DateTimeFormat("de-DE", { weekday: "short" });
