+++ src/pages/Calendar.tsx
@@ -1,0 +1,59 @@
+/*
+License:
+
+This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0. If a copy of
+the MPL was not distributed with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
+*/
+
+import { useLocation } from "preact-iso";
+import Helmet from "preact-helmet";
+
+import ByMonths from "../components/Calendar/ByMonth";
+import Downloader from "../components/Download";
+import Legend from "../components/Legend";
+import {
+  shiftModelNames,
+  shiftModelText,
+  type ShiftModels,
+} from "../../lib/constants";
+import { useTodayZeroIndex } from "../../hooks/time";
+
+import style from "../components/Calendar/style.module.css";
+
+/**
+ * Route that always displays today.
+ */
+export default function ShiftModel({
+  params: { shiftModel },
+}: {
+  params: { shiftModel: ShiftModels };
+}) {
+  const { route } = useLocation();
+  if (!shiftModelNames.includes(shiftModel)) {
+    route("/", true);
+  }
+  const [year, month] = useTodayZeroIndex();
+
+  return (
+    <main className={style.main}>
+      <Helmet
+        title={`Monat ${year}-${String(month + 1).padStart(2, "0")} | ${
+          shiftModelText[shiftModel]
+        }`}
+      />
+
+      <Legend shiftKey={shiftModel} year={year} month={month} />
+
+      <ByMonths
+        key="today"
+        shiftModel={shiftModel}
+        group={0}
+        search={null}
+        year={year}
+        month={month}
+      />
+
+      <Downloader shiftModel={shiftModel} year={year} month={month + 1} />
+    </main>
+  );
+}
