+++ styles/index.css
@@ -66,7 +66,7 @@ body {
 }
 
 #app {
-  height: 100dvh;
+  min-height: 100dvh;
 }
 
 td,
